feat: Implement CLI scripting variables, built-in commands, traps, and UI features including aliases, bookmarks, and history.
