+++ src/cli/CLIcore/CLIcore/CLIcore_UI.c
@@ -21,6 +21,7 @@
 #include "CLIcore.h"
 #include "CLIcore/cli_calc_parser.h"
 #include "CLIcore_script.h"
+#include "CLIcore_UI.h"
 
 #include <glob.h>
 #include <sys/wait.h>
@@ -57,7 +58,7 @@ extern int  yylex_destroy(void);
  * Number of ghost chars rendered on current line.
  * Set by print_ghost(), read by cli_accept_line().
  */
-static int ghost_chars_on_line = 0;
+int ghost_chars_on_line = 0;
 
 /**
  * @brief Custom accept-line handler for readline
@@ -66,7 +67,7 @@ static int ghost_chars_on_line = 0;
  * text with spaces before accepting the line, so
  * the terminal scrollback entry is clean.
  */
-static int cli_accept_line(
+int cli_accept_line(
     int count,
     int key
 )
@@ -149,25 +150,6 @@ void rl_cb_linehandler(char *linein)
 }
 #endif
 
-
-static const char *CLI_history_file(void)
-{
-    static char path[1024] = {0};
-    if(path[0] == '\0')
-    {
-        const char *home = getenv("HOME");
-        if(home)
-        {
-            snprintf(path, sizeof(path), "%s/.milk_history", home);
-        }
-        else
-        {
-            snprintf(path, sizeof(path), ".milk_history");
-        }
-    }
-    return path;
-}
-
 errno_t runCLI_prompt(char *promptstring, char *prompt)
 {
     //int color_cyan = 36;
@@ -198,7 +180,7 @@ errno_t runCLI_prompt(char *promptstring, char *prompt)
 }
 
 
-static void *xmalloc(int size)
+void *xmalloc(int size)
 {
     void *buf;
 
@@ -212,7 +194,7 @@ static void *xmalloc(int size)
     return buf;
 }
 
-static char *dupstr(char *s)
+char *dupstr(char *s)
 {
     char *r;
 
@@ -222,7 +204,7 @@ static char *dupstr(char *s)
 }
 
 #ifdef USE_READLINE
-static int levenshtein_distance(const char *s1, const char *s2)
+int levenshtein_distance(const char *s1, const char *s2)
 {
     unsigned int len1 = strlen(s1);
     unsigned int len2 = strlen(s2);
@@ -252,9 +234,9 @@ static int levenshtein_distance(const char *s1, const char *s2)
  * After a normal prefix-match pass, if nothing matched
  * and fuzzy is enabled, we restart with substring match.
  */
-static int generator_fuzzy_pass = 0;
+int generator_fuzzy_pass = 0;
 
-static char *CLI_generator(const char *text, int state)
+char *CLI_generator(const char *text, int state)
 {
     static unsigned int list_index;
     static unsigned int list_index1;
@@ -778,670 +760,7 @@ errno_t write_tracedebugfile()
     return RETURN_SUCCESS;
 }
 
-
-/*
- * ============================================================
- *  Command Alias Subsystem
- * ============================================================
- *
- * Aliases are stored in data.alias[] and persisted
- * to ~/.milk_aliases (one "name=command" per line).
- */
-
-/** @brief Path to alias persistence file */
-static const char *CLI_alias_file(void)
-{
-    static char path[1024] = {0};
-    if(path[0] == '\0')
-    {
-        const char *home = getenv("HOME");
-        if(home)
-        {
-            snprintf(path, sizeof(path),
-                     "%s/.milk_aliases", home);
-        }
-        else
-        {
-            snprintf(path, sizeof(path),
-                     ".milk_aliases");
-        }
-    }
-    return path;
-}
-
-/**
- * @brief Load aliases from ~/.milk_aliases
- */
-void cli_alias_load(void)
-{
-    data.NBalias = 0;
-    FILE *fp = fopen(CLI_alias_file(), "r");
-    if(fp == NULL)
-    {
-        return;
-    }
-
-    char line[CLI_ALIAS_NAMELEN + CLI_ALIAS_CMDLEN + 4];
-    while(fgets(line, (int) sizeof(line), fp)
-            != NULL)
-    {
-        if(data.NBalias >= CLI_MAX_ALIASES)
-        {
-            break;
-        }
-        /* Strip trailing newline */
-        line[strcspn(line, "\n")] = '\0';
-        /* Skip empty / comment lines */
-        if(line[0] == '\0' || line[0] == '#')
-        {
-            continue;
-        }
-        /* Split on first '=' */
-        char *eq = strchr(line, '=');
-        if(eq == NULL)
-        {
-            continue;
-        }
-        *eq = '\0';
-        strncpy(data.alias[data.NBalias].name,
-                line,
-                CLI_ALIAS_NAMELEN - 1);
-        data.alias[data.NBalias].name[
-            CLI_ALIAS_NAMELEN - 1] = '\0';
-        strncpy(data.alias[data.NBalias].cmd,
-                eq + 1,
-                CLI_ALIAS_CMDLEN - 1);
-        data.alias[data.NBalias].cmd[
-            CLI_ALIAS_CMDLEN - 1] = '\0';
-        data.NBalias++;
-    }
-    fclose(fp);
-}
-
-/**
- * @brief Save aliases to ~/.milk_aliases
- */
-static void cli_alias_save(void)
-{
-    FILE *fp = fopen(CLI_alias_file(), "w");
-    if(fp == NULL)
-    {
-        printf("ERROR: cannot write %s\n",
-               CLI_alias_file());
-        return;
-    }
-    for(int i = 0; i < data.NBalias; i++)
-    {
-        fprintf(fp, "%s=%s\n",
-                data.alias[i].name,
-                data.alias[i].cmd);
-    }
-    fclose(fp);
-}
-
-/**
- * @brief CLI handler: alias <name> <command...>
- *
- * Creates or updates a command alias.
- */
-errno_t cli_alias_add(void)
-{
-    if(data.cmdNBarg < 3)
-    {
-        printf("Usage: alias <name> <command...>\n");
-        return RETURN_FAILURE;
-    }
-
-    const char *name =
-        data.cmdargtoken[1].val.string;
-
-    /* Build command from args 2..N */
-    char cmd[CLI_ALIAS_CMDLEN];
-    cmd[0] = '\0';
-    for(long a = 2; a < data.cmdNBarg; a++)
-    {
-        if(a > 2)
-        {
-            strncat(cmd, " ",
-                    CLI_ALIAS_CMDLEN
-                    - strlen(cmd) - 1);
-        }
-        strncat(cmd,
-                data.cmdargtoken[a].val.string,
-                CLI_ALIAS_CMDLEN
-                - strlen(cmd) - 1);
-    }
-
-    /* Check if alias already exists — update */
-    for(int i = 0; i < data.NBalias; i++)
-    {
-        if(strcmp(data.alias[i].name, name) == 0)
-        {
-            strncpy(data.alias[i].cmd, cmd,
-                    CLI_ALIAS_CMDLEN - 1);
-            data.alias[i].cmd[
-                CLI_ALIAS_CMDLEN - 1] = '\0';
-            cli_alias_save();
-            printf("Alias updated: %s = %s\n",
-                   name, cmd);
-            return RETURN_SUCCESS;
-        }
-    }
-
-    /* Add new alias */
-    if(data.NBalias >= CLI_MAX_ALIASES)
-    {
-        printf("ERROR: alias table full (%d)\n",
-               CLI_MAX_ALIASES);
-        return RETURN_FAILURE;
-    }
-
-    strncpy(data.alias[data.NBalias].name,
-            name, CLI_ALIAS_NAMELEN - 1);
-    data.alias[data.NBalias].name[
-        CLI_ALIAS_NAMELEN - 1] = '\0';
-    strncpy(data.alias[data.NBalias].cmd,
-            cmd, CLI_ALIAS_CMDLEN - 1);
-    data.alias[data.NBalias].cmd[
-        CLI_ALIAS_CMDLEN - 1] = '\0';
-    data.NBalias++;
-
-    cli_alias_save();
-    printf("Alias created: %s = %s\n",
-           name, cmd);
-
-    return RETURN_SUCCESS;
-}
-
-/**
- * @brief CLI handler: unalias <name>
- */
-errno_t cli_alias_remove(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: unalias <name>\n");
-        return RETURN_FAILURE;
-    }
-
-    const char *name =
-        data.cmdargtoken[1].val.string;
-
-    for(int i = 0; i < data.NBalias; i++)
-    {
-        if(strcmp(data.alias[i].name, name) == 0)
-        {
-            /* Shift remaining entries */
-            for(int j = i;
-                    j < data.NBalias - 1; j++)
-            {
-                data.alias[j] = data.alias[j + 1];
-            }
-            data.NBalias--;
-            cli_alias_save();
-            printf("Alias removed: %s\n", name);
-            return RETURN_SUCCESS;
-        }
-    }
-
-    printf("Alias '%s' not found\n", name);
-    return RETURN_FAILURE;
-}
-
-/**
- * @brief CLI handler: aliases
- */
-errno_t cli_alias_list(void)
-{
-    if(data.NBalias == 0)
-    {
-        printf("No aliases defined.\n");
-        return RETURN_SUCCESS;
-    }
-    printf("--- Aliases (%d) ---\n",
-           data.NBalias);
-    for(int i = 0; i < data.NBalias; i++)
-    {
-        printf("  %-16s = %s\n",
-               data.alias[i].name,
-               data.alias[i].cmd);
-    }
-    return RETURN_SUCCESS;
-}
-
-/**
- * @brief Expand alias in data.CLIcmdline
- *
- * If the first word matches an alias name,
- * substitute it with the alias command, keeping
- * any trailing arguments.
- */
-static void cli_alias_expand(void)
-{
-    if(data.NBalias == 0)
-    {
-        return;
-    }
-
-    /* Extract first word */
-    char firstword[CLI_ALIAS_NAMELEN];
-    int  fwlen = 0;
-    const char *p = data.CLIcmdline;
-
-    /* Skip leading whitespace */
-    while(*p == ' ' || *p == '\t')
-    {
-        p++;
-    }
-    while(*p != '\0' && *p != ' '
-            && *p != '\t' && *p != '\n')
-    {
-        if(fwlen < CLI_ALIAS_NAMELEN - 1)
-        {
-            firstword[fwlen++] = *p;
-        }
-        p++;
-    }
-    firstword[fwlen] = '\0';
-
-    if(fwlen == 0)
-    {
-        return;
-    }
-
-    /* Search aliases */
-    for(int i = 0; i < data.NBalias; i++)
-    {
-        if(strcmp(data.alias[i].name,
-                 firstword) == 0)
-        {
-            /* Build expanded line */
-            char expanded[STRINGMAXLEN_CLICMDLINE];
-            snprintf(expanded,
-                     STRINGMAXLEN_CLICMDLINE,
-                     "%s%s",
-                     data.alias[i].cmd,
-                     p); /* p points to rest */
-            strncpy(data.CLIcmdline, expanded,
-                    STRINGMAXLEN_CLICMDLINE - 1);
-            data.CLIcmdline[
-                STRINGMAXLEN_CLICMDLINE - 1]
-                = '\0';
-            return;
-        }
-    }
-}
-
-
-/*
- * ============================================================
- *  Watch Command
- * ============================================================
- */
-
-/**
- * @brief CLI handler: watch <interval_ms> <command>
- *
- * Repeats a command at a fixed interval with
- * in-place terminal refresh. Press any key to stop.
- */
-errno_t cli_watch(void)
-{
-    if(data.cmdNBarg < 3)
-    {
-        printf(
-            "Usage: watch <interval_ms>"
-            " <command...>\n");
-        return RETURN_FAILURE;
-    }
-
-    long interval_ms =
-        data.cmdargtoken[1].val.numl;
-    if(interval_ms < 10)
-    {
-        interval_ms = 10;
-    }
-
-    /* Build command from remaining args */
-    char watchcmd[STRINGMAXLEN_CLICMDLINE];
-    watchcmd[0] = '\0';
-    for(long a = 2; a < data.cmdNBarg; a++)
-    {
-        if(a > 2)
-        {
-            strncat(watchcmd, " ",
-                    STRINGMAXLEN_CLICMDLINE
-                    - strlen(watchcmd) - 1);
-        }
-        strncat(watchcmd,
-                data.cmdargtoken[a].val.string,
-                STRINGMAXLEN_CLICMDLINE
-                - strlen(watchcmd) - 1);
-    }
-
-    /* Switch terminal to raw mode so we can
-     * detect single keypresses without Enter.
-     * Readline leaves the terminal in cooked
-     * mode which buffers input. */
-    struct termios orig_termios;
-    struct termios raw_termios;
-    tcgetattr(STDIN_FILENO, &orig_termios);
-    raw_termios = orig_termios;
-    raw_termios.c_lflag &=
-        ~((tcflag_t) ICANON | (tcflag_t) ECHO);
-    raw_termios.c_cc[VMIN]  = 0;
-    raw_termios.c_cc[VTIME] = 0;
-    tcsetattr(STDIN_FILENO, TCSANOW,
-              &raw_termios);
-
-    /* Loop until keypress */
-    for(;;)
-    {
-        /* Clear screen, move cursor to top */
-        printf("\033[2J\033[H");
-
-        /* Print header */
-        {
-            time_t now = time(NULL);
-            struct tm *tm = localtime(&now);
-            printf(
-                "Every %ldms: %s   "
-                "%02d:%02d:%02d"
-                "  (press any key to stop)\n\n",
-                interval_ms, watchcmd,
-                tm->tm_hour, tm->tm_min,
-                tm->tm_sec);
-        }
-
-        /* Execute the command */
-        strncpy(data.CLIcmdline, watchcmd,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        data.CLIcmdline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-        CLI_execute_line();
-
-        fflush(stdout);
-
-        /* Sleep in small increments, checking
-         * for keypress */
-        {
-            long slept = 0;
-            long step = 50000; /* 50 ms */
-            while(slept < interval_ms * 1000)
-            {
-                struct timeval tv;
-                fd_set fds;
-                tv.tv_sec  = 0;
-                tv.tv_usec = step;
-                FD_ZERO(&fds);
-                FD_SET(STDIN_FILENO, &fds);
-                int r = select(
-                    STDIN_FILENO + 1,
-                    &fds, NULL, NULL, &tv);
-                if(r > 0)
-                {
-                    /* Consume the keypress */
-                    char discard;
-                    if(read(STDIN_FILENO,
-                            &discard, 1) > 0)
-                    {
-                        /* ignore value */
-                    }
-                    goto watch_done;
-                }
-                slept += step;
-            }
-        }
-    }
-
-watch_done:
-    /* Restore original terminal settings */
-    tcsetattr(STDIN_FILENO, TCSANOW,
-              &orig_termios);
-    printf("\nwatch stopped.\n");
-
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Startup Script (~/.milkrc)
- * ============================================================
- */
-
-void cli_milkrc_load(void)
-{
-    char rcpath[STRINGMAXLEN_FULLFILENAME];
-    snprintf(rcpath,
-             STRINGMAXLEN_FULLFILENAME,
-             "%s/.milkrc", getenv("HOME"));
-
-    FILE *fp = fopen(rcpath, "r");
-    if(fp == NULL)
-    {
-        return;
-    }
-
-    char line[STRINGMAXLEN_CLICMDLINE];
-    while(fgets(line, STRINGMAXLEN_CLICMDLINE,
-                fp) != NULL)
-    {
-        size_t len = strlen(line);
-        if(len > 0 && line[len - 1] == '\n')
-        {
-            line[len - 1] = '\0';
-        }
-        const char *p = line;
-        while(*p == ' ' || *p == '\t')
-        {
-            p++;
-        }
-        if(*p == '\0' || *p == '#')
-        {
-            continue;
-        }
-        strncpy(data.CLIcmdline, line,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        data.CLIcmdline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-        CLI_execute_line();
-    }
-    fclose(fp);
-}
-
-
-/*
- * ============================================================
- *  Command Timing
- * ============================================================
- */
-
-errno_t cli_time(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: time <command...>\n");
-        return RETURN_FAILURE;
-    }
-    char timecmd[STRINGMAXLEN_CLICMDLINE];
-    timecmd[0] = '\0';
-    for(long a = 1; a < data.cmdNBarg; a++)
-    {
-        if(a > 1)
-        {
-            strncat(timecmd, " ",
-                    STRINGMAXLEN_CLICMDLINE
-                    - strlen(timecmd) - 1);
-        }
-        strncat(timecmd,
-                data.cmdargtoken[a].val.string,
-                STRINGMAXLEN_CLICMDLINE
-                - strlen(timecmd) - 1);
-    }
-    struct timespec t0;
-    struct timespec t1;
-    clock_gettime(CLOCK_MONOTONIC, &t0);
-
-    strncpy(data.CLIcmdline, timecmd,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    data.CLIcmdline[
-        STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-    CLI_execute_line();
-
-    clock_gettime(CLOCK_MONOTONIC, &t1);
-    {
-        double elapsed =
-            (double)(t1.tv_sec - t0.tv_sec)
-            + 1.0e-9
-            * (double)(t1.tv_nsec - t0.tv_nsec);
-        printf(
-            "\n\033[33mElapsed: %.6f s\033[0m\n",
-            elapsed);
-    }
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Command Statistics
- * ============================================================
- */
-
-errno_t cli_cmdstats(void)
-{
-    typedef struct
-    {
-        const char *key;
-        uint32_t    count;
-    } CmdStatEntry;
-
-    CmdStatEntry entries[DATA_NB_MAX_COMMAND];
-    int nused = 0;
-    for(uint32_t i = 0; i < data.NBcmd; i++)
-    {
-        if(data.cmd[i].callcount > 0)
-        {
-            entries[nused].key =
-                data.cmd[i].key;
-            entries[nused].count =
-                data.cmd[i].callcount;
-            nused++;
-        }
-    }
-    if(nused == 0)
-    {
-        printf("No commands executed yet.\n");
-        return RETURN_SUCCESS;
-    }
-    for(int i = 1; i < nused; i++)
-    {
-        CmdStatEntry tmp = entries[i];
-        int j = i - 1;
-        while(j >= 0
-                && entries[j].count < tmp.count)
-        {
-            entries[j + 1] = entries[j];
-            j--;
-        }
-        entries[j + 1] = tmp;
-    }
-    int show = nused < 20 ? nused : 20;
-    printf("\n\033[1mCommand usage "
-           "(top %d):\033[0m\n", show);
-    printf("  %-30s  %s\n", "COMMAND", "CALLS");
-    printf("  %-30s  %s\n",
-           "------------------------------",
-           "-----");
-    for(int i = 0; i < show; i++)
-    {
-        printf("  %-30s  %u\n",
-               entries[i].key,
-               entries[i].count);
-    }
-    printf("\n");
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Syntax Highlighting (optional, ON by default)
- * ============================================================
- */
-
-#ifdef USE_READLINE
-
-errno_t cli_timing_toggle(void)
-{
-    if(data.cmdNBarg >= 2)
-    {
-        const char *arg =
-            data.cmdargtoken[1].val.string;
-        if(strcmp(arg, "on") == 0
-                || strcmp(arg, "1") == 0)
-        {
-            data.print_cmd_timing = 1;
-            printf("Command execution timing ON\n");
-        }
-        else if(strcmp(arg, "off") == 0
-                || strcmp(arg, "0") == 0)
-        {
-            data.print_cmd_timing = 0;
-            printf("Command execution timing OFF\n");
-        }
-        else
-        {
-            printf("Usage: cli.timing [on|off]\n");
-        }
-    }
-    else
-    {
-        data.print_cmd_timing =
-            !data.print_cmd_timing;
-        printf("Command execution timing %s\n",
-               data.print_cmd_timing
-               ? "ON" : "OFF");
-    }
-    return RETURN_SUCCESS;
-}
-
-errno_t cli_syntax_highlight_toggle(void)
-{
-    if(data.cmdNBarg >= 2)
-    {
-        const char *arg =
-            data.cmdargtoken[1].val.string;
-        if(strcmp(arg, "on") == 0
-                || strcmp(arg, "1") == 0)
-        {
-            data.syntax_highlight = 1;
-            printf("Syntax highlighting ON\n");
-        }
-        else if(strcmp(arg, "off") == 0
-                || strcmp(arg, "0") == 0)
-        {
-            data.syntax_highlight = 0;
-            printf("Syntax highlighting OFF\n");
-        }
-        else
-        {
-            printf("Usage: synhl [on|off]\n");
-        }
-    }
-    else
-    {
-        data.syntax_highlight =
-            !data.syntax_highlight;
-        printf("Syntax highlighting %s\n",
-               data.syntax_highlight
-               ? "ON" : "OFF");
-    }
-    return RETURN_SUCCESS;
-}
-
-static int cli_is_command(const char *word)
+int cli_is_command(const char *word)
 {
     for(uint32_t i = 0; i < data.NBcmd; i++)
     {
@@ -1453,7 +772,7 @@ static int cli_is_command(const char *word)
     return 0;
 }
 
-static void cli_highlight_redisplay(void)
+void cli_highlight_redisplay(void)
 {
     if(!data.syntax_highlight)
     {
@@ -1533,868 +852,6 @@ static void cli_highlight_redisplay(void)
     fflush(rl_outstream);
 }
 
-#endif /* USE_READLINE */
-
-
-/*
- * ============================================================
- *  Persistent History (~/.milk_history)
- * ============================================================
- */
-
-#define MILK_HISTORY_MAXLINES 1000
-
-void cli_history_load(void)
-{
-#ifdef USE_READLINE
-    char hpath[STRINGMAXLEN_FULLFILENAME];
-    snprintf(hpath,
-             STRINGMAXLEN_FULLFILENAME,
-             "%s/.milk_history",
-             getenv("HOME"));
-    read_history(hpath);
-#endif
-}
-
-void cli_history_save(void)
-{
-#ifdef USE_READLINE
-    char hpath[STRINGMAXLEN_FULLFILENAME];
-    snprintf(hpath,
-             STRINGMAXLEN_FULLFILENAME,
-             "%s/.milk_history",
-             getenv("HOME"));
-    write_history(hpath);
-    history_truncate_file(hpath,
-                          MILK_HISTORY_MAXLINES);
-#endif
-}
-
-
-/*
- * ============================================================
- *  Structured History Log (~/.milk_history_log)
- *
- *  Append-only TSV file that records every
- *  command with session metadata:
- *    <timestamp>\t<session_id>\t<tty>\t<cmd>
- * ============================================================
- */
-
-/**
- * @brief Path to structured history log file
- */
-static const char *CLI_history_log_file(void)
-{
-    static char path[1024] = {0};
-    if(path[0] == '\0')
-    {
-        const char *home = getenv("HOME");
-        if(home)
-        {
-            snprintf(path, sizeof(path),
-                     "%s/.milk_history_log",
-                     home);
-        }
-        else
-        {
-            snprintf(path, sizeof(path),
-                     ".milk_history_log");
-        }
-    }
-    return path;
-}
-
-/**
- * @brief Initialize session identity
- *
- * Records processname, PID, terminal,
- * and start time into data struct.
- * Called once at startup.
- */
-void cli_history_log_init(void)
-{
-    pid_t pid = getpid();
-    snprintf(data.session_id,
-             sizeof(data.session_id),
-             "%s-%d",
-             data.processname, (int) pid);
-
-    {
-        const char *tty = ttyname(STDIN_FILENO);
-        if(tty)
-        {
-            strncpy(data.session_tty, tty,
-                    sizeof(data.session_tty) - 1);
-            data.session_tty[
-                sizeof(data.session_tty) - 1]
-                = '\0';
-        }
-        else
-        {
-            strncpy(data.session_tty, "?",
-                    sizeof(data.session_tty) - 1);
-        }
-    }
-
-    clock_gettime(CLOCK_REALTIME,
-                  &data.session_start);
-}
-
-/**
- * @brief Log command to structured history file
- *
- * Appends a TSV line:
- *   <ISO-timestamp>\t<session_id>\t<tty>\t<cmd>
- *
- * Called alongside add_history() in the main
- * command dispatch loop.
- */
-static void cli_history_log_cmd(
-    const char *cmd
-)
-{
-    if(cmd == NULL || cmd[0] == '\0')
-    {
-        return;
-    }
-
-    FILE *fp = fopen(CLI_history_log_file(), "a");
-    if(fp == NULL)
-    {
-        return;
-    }
-
-    {
-        time_t now = time(NULL);
-        char tbuf[64];
-        strftime(tbuf, sizeof(tbuf),
-                 "%Y-%m-%dT%H:%M:%S",
-                 localtime(&now));
-        fprintf(fp, "%s\t%s\t%s\t%s\n",
-                tbuf,
-                data.session_id,
-                data.session_tty,
-                cmd);
-    }
-    fclose(fp);
-}
-
-
-/**
- * @brief Helper: read history log and display
- *
- * @param filter_session  if non-NULL, only show
- *        entries matching this session_id
- * @param max_entries     show at most this many
- *        entries (0 = all)
- */
-static void history_log_display(
-    const char *filter_session,
-    int         max_entries
-)
-{
-    FILE *fp = fopen(CLI_history_log_file(), "r");
-    if(fp == NULL)
-    {
-        printf("No history log found (%s)\n",
-               CLI_history_log_file());
-        return;
-    }
-
-    /* First pass: count matching lines */
-    char line[2048];
-    int total = 0;
-
-    /* Store lines in a simple dynamic array */
-    int cap = 1024;
-    char **lines = (char **) malloc(
-        (size_t) cap * sizeof(char *));
-    if(lines == NULL)
-    {
-        fclose(fp);
-        printf("Memory allocation error\n");
-        return;
-    }
-
-    while(fgets(line, (int) sizeof(line), fp))
-    {
-        /* Remove trailing newline */
-        {
-            size_t len = strlen(line);
-            if(len > 0
-               && line[len - 1] == '\n')
-            {
-                line[len - 1] = '\0';
-            }
-        }
-
-        /* Parse: timestamp\tsession\ttty\tcmd */
-        if(filter_session != NULL)
-        {
-            char *tab1 = strchr(line, '\t');
-            if(tab1 == NULL)
-            {
-                continue;
-            }
-            char *tab2 = strchr(tab1 + 1, '\t');
-            if(tab2 == NULL)
-            {
-                continue;
-            }
-            int slen = (int)(tab2 - tab1 - 1);
-            if(slen
-               != (int) strlen(filter_session)
-               || strncmp(tab1 + 1,
-                          filter_session,
-                          (size_t) slen) != 0)
-            {
-                continue;
-            }
-        }
-
-        /* Store line */
-        if(total >= cap)
-        {
-            cap *= 2;
-            char **tmp = (char **) realloc(
-                lines,
-                (size_t) cap * sizeof(char *));
-            if(tmp == NULL)
-            {
-                break;
-            }
-            lines = tmp;
-        }
-        lines[total] = strdup(line);
-        total++;
-    }
-    fclose(fp);
-
-    if(total == 0)
-    {
-        if(filter_session)
-        {
-            printf("No history for session"
-                   " '%s'\n", filter_session);
-        }
-        else
-        {
-            printf("No history entries\n");
-        }
-        free(lines);
-        return;
-    }
-
-    /* Determine start index */
-    int start = 0;
-    if(max_entries > 0
-       && max_entries < total)
-    {
-        start = total - max_entries;
-    }
-
-    /* Print header */
-    if(filter_session != NULL)
-    {
-        printf("\033[1;36m %-19s  %s\033[0m\n",
-               "Time", "Command");
-    }
-    else
-    {
-        printf("\033[1;36m %-24s %-19s  "
-               "%s\033[0m\n",
-               "Session",
-               "Time", "Command");
-    }
-
-    /* Print entries */
-    for(int i = start; i < total; i++)
-    {
-        /* Parse fields */
-        char *ts = lines[i];
-        char *tab1 = strchr(ts, '\t');
-        if(tab1 == NULL)
-        {
-            continue;
-        }
-        *tab1 = '\0';
-        char *sid = tab1 + 1;
-        char *tab2 = strchr(sid, '\t');
-        if(tab2 == NULL)
-        {
-            continue;
-        }
-        *tab2 = '\0';
-        char *tty = tab2 + 1;
-        char *tab3 = strchr(tty, '\t');
-        if(tab3 == NULL)
-        {
-            continue;
-        }
-        *tab3 = '\0';
-        char *cmd = tab3 + 1;
-
-        /* Shorten tty: /dev/pts/3 → /pts/3 */
-        const char *tty_short = tty;
-        if(strncmp(tty, "/dev", 4) == 0)
-        {
-            tty_short = tty + 4;
-        }
-
-        if(filter_session != NULL)
-        {
-            /* local: no session column */
-            printf(" %-19s  %s\n", ts, cmd);
-        }
-        else
-        {
-            /* global: show session + tty */
-            char sess_col[40];
-            snprintf(sess_col,
-                     sizeof(sess_col),
-                     "%s %s", sid, tty_short);
-            printf(" %-24s %-19s  %s\n",
-                   sess_col, ts, cmd);
-        }
-    }
-
-    /* Summary line */
-    int shown = total - start;
-    if(filter_session)
-    {
-        printf("(%d entr%s, session %s)\n",
-               shown,
-               shown == 1 ? "y" : "ies",
-               filter_session);
-    }
-    else
-    {
-        printf("(%d entr%s)\n",
-               shown,
-               shown == 1 ? "y" : "ies");
-    }
-
-    /* Free */
-    for(int i = 0; i < total; i++)
-    {
-        free(lines[i]);
-    }
-    free(lines);
-}
-
-/**
- * @brief Global history command
- *
- * Usage:
- *   ghistory         — last 20 entries
- *   ghistory N       — last N entries
- *   ghistory -s SID  — filter by session ID
- */
-errno_t cli_ghistory(void)
-{
-    int n = 20;
-    const char *filter = NULL;
-
-    if(data.cmdNBarg >= 2)
-    {
-        const char *arg1 =
-            data.cmdargtoken[1].val.string;
-        if(strcmp(arg1, "-s") == 0)
-        {
-            if(data.cmdNBarg >= 3)
-            {
-                filter =
-                    data.cmdargtoken[2]
-                        .val.string;
-            }
-            else
-            {
-                printf("Usage: ghistory"
-                       " -s <session_id>\n");
-                return RETURN_FAILURE;
-            }
-            n = 0; /* show all for session */
-        }
-        else
-        {
-            n = atoi(arg1);
-            if(n <= 0)
-            {
-                n = 20;
-            }
-        }
-    }
-
-    history_log_display(filter, n);
-    return RETURN_SUCCESS;
-}
-
-/**
- * @brief Local (session) history command
- *
- * Usage:
- *   lhistory         — all entries, this session
- *   lhistory N       — last N entries
- */
-errno_t cli_lhistory(void)
-{
-    int n = 0; /* default: show all */
-
-    if(data.cmdNBarg >= 2)
-    {
-        n = atoi(
-            data.cmdargtoken[1].val.string);
-        if(n <= 0)
-        {
-            n = 0;
-        }
-    }
-
-    history_log_display(data.session_id, n);
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  History Expansion (!! and !$)
- * ============================================================
- *
- * Called at the very start of CLI_execute_line(),
- * before alias expansion.
- *
- * !!    → replace with last executed command
- * !$    → replace with last argument of previous cmd
- * !<prefix> → last command starting with <prefix>
- */
-
-static void cli_history_expand(void)
-{
-#ifdef USE_READLINE
-    char *line = data.CLIcmdline;
-
-    /* Quick check: must start with '!' */
-    if(line[0] != '!')
-    {
-        return;
-    }
-
-    /* !! — replay last command */
-    if(line[1] == '!')
-    {
-        HIST_ENTRY *prev = history_get(
-            history_length);
-        if(prev != NULL)
-        {
-            char suffix[STRINGMAXLEN_CLICMDLINE];
-            suffix[0] = '\0';
-            if(line[2] != '\0')
-            {
-                strncpy(suffix, line + 2,
-                        STRINGMAXLEN_CLICMDLINE
-                        - 1);
-                suffix[
-                    STRINGMAXLEN_CLICMDLINE - 1]
-                    = '\0';
-            }
-            snprintf(data.CLIcmdline,
-                     STRINGMAXLEN_CLICMDLINE,
-                     "%s%s",
-                     prev->line, suffix);
-            printf(">> %s\n", data.CLIcmdline);
-        }
-        return;
-    }
-
-    /* !$ — last argument of previous command */
-    if(line[1] == '$')
-    {
-        if(data.last_argument[0] != '\0')
-        {
-            char rest[STRINGMAXLEN_CLICMDLINE];
-            strncpy(rest, line + 2,
-                    STRINGMAXLEN_CLICMDLINE - 1);
-            rest[
-                STRINGMAXLEN_CLICMDLINE - 1]
-                = '\0';
-            snprintf(data.CLIcmdline,
-                     STRINGMAXLEN_CLICMDLINE,
-                     "%s%s",
-                     data.last_argument, rest);
-            printf(">> %s\n", data.CLIcmdline);
-        }
-        return;
-    }
-
-    /* !<prefix> — last command starting with it */
-    {
-        const char *prefix = line + 1;
-        size_t plen = strlen(prefix);
-        /* Trim trailing spaces from prefix */
-        while(plen > 0
-                && prefix[plen - 1] == ' ')
-        {
-            plen--;
-        }
-        if(plen == 0)
-        {
-            return;
-        }
-        HIST_ENTRY **hist = history_list();
-        if(hist == NULL)
-        {
-            return;
-        }
-        int hlen = history_length;
-        for(int i = hlen - 1; i >= 0; i--)
-        {
-            if(strncmp(hist[i]->line,
-                       prefix, plen) == 0)
-            {
-                strncpy(data.CLIcmdline,
-                        hist[i]->line,
-                        STRINGMAXLEN_CLICMDLINE
-                        - 1);
-                data.CLIcmdline[
-                    STRINGMAXLEN_CLICMDLINE - 1]
-                    = '\0';
-                printf(">> %s\n",
-                       data.CLIcmdline);
-                return;
-            }
-        }
-        printf("!%.*s: event not found\n",
-               (int) plen, prefix);
-        data.CLIcmdline[0] = '\0';
-    }
-#endif
-}
-
-
-/**
- * @brief Save last argument after command execution
- */
-static void cli_save_last_argument(void)
-{
-    if(data.cmdNBarg > 1)
-    {
-        long last = data.cmdNBarg - 1;
-        strncpy(data.last_argument,
-                data.cmdargtoken[last].val.string,
-                sizeof(data.last_argument) - 1);
-        data.last_argument[
-            sizeof(data.last_argument) - 1]
-            = '\0';
-    }
-}
-
-
-/*
- * ============================================================
- *  Source Command — execute a milk script file
- * ============================================================
- */
-
-errno_t cli_source(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: source <filename>\n");
-        return RETURN_FAILURE;
-    }
-    const char *fname =
-        data.cmdargtoken[1].val.string;
-    FILE *fp = fopen(fname, "r");
-    if(fp == NULL)
-    {
-        printf("source: cannot open '%s'\n",
-               fname);
-        return RETURN_FAILURE;
-    }
-    char line[STRINGMAXLEN_CLICMDLINE];
-    int lineno = 0;
-    while(fgets(line, STRINGMAXLEN_CLICMDLINE,
-                fp) != NULL)
-    {
-        lineno++;
-        {
-            size_t len = strlen(line);
-            if(len > 0
-                    && line[len - 1] == '\n')
-            {
-                line[len - 1] = '\0';
-            }
-        }
-        const char *p = line;
-        while(*p == ' ' || *p == '\t')
-        {
-            p++;
-        }
-        if(*p == '\0' || *p == '#')
-        {
-            continue;
-        }
-        strncpy(data.CLIcmdline, line,
-                STRINGMAXLEN_CLICMDLINE - 1);
-        data.CLIcmdline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-        errno_t ret = CLI_execute_line();
-        if(ret != RETURN_SUCCESS)
-        {
-            printf(
-                "\033[31m[source:%s:%d] "
-                "error\033[0m\n",
-                fname, lineno);
-        }
-    }
-    fclose(fp);
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Save Script — export variables and functions
- * ============================================================
- */
-
-/**
- * @brief Write all CLI variables and user functions
- *        to a file that can be sourced later.
- *
- * Usage: savescript <filename>
- */
-errno_t cli_savescript(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: savescript <filename>\n");
-        return RETURN_FAILURE;
-    }
-    const char *fname =
-        data.cmdargtoken[1].val.string;
-    FILE *fp = fopen(fname, "w");
-    if(fp == NULL)
-    {
-        printf("savescript: cannot open "
-               "'%s' for writing\n", fname);
-        return RETURN_FAILURE;
-    }
-
-    fprintf(fp, "# milk-cli script\n");
-    fprintf(fp,
-            "# saved by savescript command\n\n");
-
-    /* Export variables */
-    int nv = 0;
-    for(int i = 0; i < CLI_MAX_VARS; i++)
-    {
-        if(cli_vars[i].used)
-        {
-            fprintf(fp, "%s=%s\n",
-                    cli_vars[i].name,
-                    cli_vars[i].val);
-            nv++;
-        }
-    }
-    if(nv > 0)
-    {
-        fprintf(fp, "\n");
-    }
-
-    /* Export user-defined functions */
-    int nf = 0;
-    for(int i = 0; i < CLI_MAX_FUNCS; i++)
-    {
-        if(cli_funcs[i].used)
-        {
-            fprintf(fp, "function %s {\n",
-                    cli_funcs[i].name);
-            for(int j = 0;
-                j < cli_funcs[i].nbody; j++)
-            {
-                fprintf(fp, "%s\n",
-                        cli_funcs[i].body[j]);
-            }
-            fprintf(fp, "}\n\n");
-            nf++;
-        }
-    }
-
-    fclose(fp);
-    printf("Saved %d variables, %d functions "
-           "to '%s'\n", nv, nf, fname);
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Save History — export readline history
- * ============================================================
- */
-
-/**
- * @brief Write readline command history to a file.
- *
- * Usage: savehistory <filename>
- */
-errno_t cli_savehistory(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: savehistory "
-               "<filename>\n");
-        return RETURN_FAILURE;
-    }
-    const char *fname =
-        data.cmdargtoken[1].val.string;
-
-#ifdef USE_READLINE
-    if(write_history(fname) != 0)
-    {
-        printf("savehistory: failed to write "
-               "'%s'\n", fname);
-        return RETURN_FAILURE;
-    }
-    printf("History saved to '%s'\n", fname);
-    return RETURN_SUCCESS;
-#else
-    printf("savehistory: readline not "
-           "available\n");
-    (void) fname;
-    return RETURN_FAILURE;
-#endif
-}
-
-
-/*
- * ============================================================
- *  Configurable Prompt — setprompt command
- * ============================================================
- *
- * Format tokens:
- *   %h = hostname
- *   %u = username
- *   %d = cwd basename
- *   %t = HH:MM:SS
- *   %n = CLI process name (data.processname)
- */
-
-/** prompt_format stored in data struct is TBD;
- *  for now use a file-scope buffer. */
-static char cli_prompt_format[200] = "";
-
-/**
- * @brief Build prompt string from format tokens
- */
-void cli_build_prompt(
-    const char *fmt,
-    char       *out,
-    int         maxlen
-)
-{
-    int pos = 0;
-    for(int i = 0; fmt[i] != '\0'
-            && pos < maxlen - 1; i++)
-    {
-        if(fmt[i] == '%' && fmt[i + 1] != '\0')
-        {
-            i++;
-            switch(fmt[i])
-            {
-            case 'h':
-            {
-                char hn[64];
-                gethostname(hn, sizeof(hn));
-                pos += snprintf(out + pos,
-                    (size_t)(maxlen - pos),
-                    "%s", hn);
-                break;
-            }
-            case 'u':
-            {
-                const char *u = getenv("USER");
-                pos += snprintf(out + pos,
-                    (size_t)(maxlen - pos),
-                    "%s", u ? u : "?");
-                break;
-            }
-            case 'd':
-            {
-                char cwd[256];
-                if(getcwd(cwd, sizeof(cwd)))
-                {
-                    char *base = strrchr(cwd,
-                                         '/');
-                    pos += snprintf(out + pos,
-                        (size_t)(maxlen - pos),
-                        "%s",
-                        base ? base + 1 : cwd);
-                }
-                break;
-            }
-            case 't':
-            {
-                time_t now = time(NULL);
-                struct tm *tm = localtime(&now);
-                pos += (int) strftime(
-                    out + pos,
-                    (size_t)(maxlen - pos),
-                    "%H:%M:%S", tm);
-                break;
-            }
-            case 'n':
-                pos += snprintf(out + pos,
-                    (size_t)(maxlen - pos),
-                    "%s", data.processname);
-                break;
-            default:
-                if(pos < maxlen - 2)
-                {
-                    out[pos++] = '%';
-                    out[pos++] = fmt[i];
-                }
-                break;
-            }
-        }
-        else
-        {
-            out[pos++] = fmt[i];
-        }
-    }
-    out[pos] = '\0';
-}
-
-errno_t cli_setprompt(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        if(cli_prompt_format[0] != '\0')
-        {
-            printf("Current prompt format: "
-                   "'%s'\n",
-                   cli_prompt_format);
-        }
-        else
-        {
-            printf("Using default prompt\n");
-        }
-        printf("Tokens: %%h=host %%u=user "
-               "%%d=dir %%t=time %%n=name\n");
-        return RETURN_SUCCESS;
-    }
-    strncpy(cli_prompt_format,
-            data.cmdargtoken[1].val.string,
-            sizeof(cli_prompt_format) - 1);
-    cli_prompt_format[
-        sizeof(cli_prompt_format) - 1] = '\0';
-    printf("Prompt set to: '%s'\n",
-           cli_prompt_format);
-    return RETURN_SUCCESS;
-}
-
 
 /*
  * ============================================================
@@ -2402,2083 +859,9 @@ errno_t cli_setprompt(void)
  * ============================================================
  */
 
-#define BOOKMARK_MAX       64
-#define BOOKMARK_NAMELEN   64
-#define BOOKMARK_CMDLEN   512
 
-struct bookmark_entry
-{
-    char name[BOOKMARK_NAMELEN];
-    char cmd[BOOKMARK_CMDLEN];
-};
 
-static struct bookmark_entry
-    bookmarks[BOOKMARK_MAX];
-static int bookmark_count = 0;
 
-/**
- * @brief Load bookmarks from ~/.milk_bookmarks
- */
-void cli_bookmark_load(void)
-{
-    char path[STRINGMAXLEN_FULLFILENAME];
-    snprintf(path,
-             STRINGMAXLEN_FULLFILENAME,
-             "%s/.milk_bookmarks",
-             getenv("HOME"));
-    FILE *fp = fopen(path, "r");
-    if(fp == NULL)
-    {
-        return;
-    }
-    bookmark_count = 0;
-    char line[1024];
-    while(fgets(line, (int) sizeof(line), fp)
-            && bookmark_count < BOOKMARK_MAX)
-    {
-        size_t len = strlen(line);
-        if(len > 0 && line[len - 1] == '\n')
-        {
-            line[len - 1] = '\0';
-        }
-        char *tab = strchr(line, '\t');
-        if(tab == NULL)
-        {
-            continue;
-        }
-        *tab = '\0';
-        strncpy(
-            bookmarks[bookmark_count].name,
-            line,
-            BOOKMARK_NAMELEN - 1);
-        bookmarks[bookmark_count].name[
-            BOOKMARK_NAMELEN - 1] = '\0';
-        strncpy(
-            bookmarks[bookmark_count].cmd,
-            tab + 1,
-            BOOKMARK_CMDLEN - 1);
-        bookmarks[bookmark_count].cmd[
-            BOOKMARK_CMDLEN - 1] = '\0';
-        bookmark_count++;
-    }
-    fclose(fp);
-}
-
-/**
- * @brief Save bookmarks to ~/.milk_bookmarks
- */
-static void cli_bookmark_save(void)
-{
-    char path[STRINGMAXLEN_FULLFILENAME];
-    snprintf(path,
-             STRINGMAXLEN_FULLFILENAME,
-             "%s/.milk_bookmarks",
-             getenv("HOME"));
-    FILE *fp = fopen(path, "w");
-    if(fp == NULL)
-    {
-        return;
-    }
-    for(int i = 0; i < bookmark_count; i++)
-    {
-        fprintf(fp, "%s\t%s\n",
-                bookmarks[i].name,
-                bookmarks[i].cmd);
-    }
-    fclose(fp);
-}
-
-errno_t cli_bookmark(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage:\n"
-               "  bookmark save <name> "
-               "\"cmd1 ; cmd2\"\n"
-               "  bookmark run  <name>\n"
-               "  bookmark list\n"
-               "  bookmark rm   <name>\n");
-        return RETURN_SUCCESS;
-    }
-    const char *action =
-        data.cmdargtoken[1].val.string;
-
-    if(strcmp(action, "list") == 0)
-    {
-        if(bookmark_count == 0)
-        {
-            printf("No bookmarks saved\n");
-        }
-        for(int i = 0; i < bookmark_count; i++)
-        {
-            printf("  \033[1m%-16s\033[0m %s\n",
-                   bookmarks[i].name,
-                   bookmarks[i].cmd);
-        }
-        return RETURN_SUCCESS;
-    }
-
-    if(strcmp(action, "save") == 0)
-    {
-        if(data.cmdNBarg < 4)
-        {
-            printf("Usage: bookmark save "
-                   "<name> \"cmd\"\n");
-            return RETURN_FAILURE;
-        }
-        if(bookmark_count >= BOOKMARK_MAX)
-        {
-            printf("Bookmark limit reached\n");
-            return RETURN_FAILURE;
-        }
-        strncpy(
-            bookmarks[bookmark_count].name,
-            data.cmdargtoken[2].val.string,
-            BOOKMARK_NAMELEN - 1);
-        bookmarks[bookmark_count].name[
-            BOOKMARK_NAMELEN - 1] = '\0';
-        /* Join remaining args as command */
-        {
-            char cmd[BOOKMARK_CMDLEN] = "";
-            for(long a = 3;
-                    a < data.cmdNBarg; a++)
-            {
-                if(a > 3)
-                {
-                    strncat(cmd, " ",
-                        BOOKMARK_CMDLEN
-                        - strlen(cmd) - 1);
-                }
-                strncat(cmd,
-                    data.cmdargtoken[a]
-                        .val.string,
-                    BOOKMARK_CMDLEN
-                    - strlen(cmd) - 1);
-            }
-            strncpy(
-                bookmarks[bookmark_count].cmd,
-                cmd, BOOKMARK_CMDLEN - 1);
-            bookmarks[bookmark_count].cmd[
-                BOOKMARK_CMDLEN - 1] = '\0';
-        }
-        bookmark_count++;
-        cli_bookmark_save();
-        printf("Bookmark '%s' saved\n",
-               data.cmdargtoken[2].val.string);
-        return RETURN_SUCCESS;
-    }
-
-    if(strcmp(action, "run") == 0)
-    {
-        if(data.cmdNBarg < 3)
-        {
-            printf("Usage: bookmark run "
-                   "<name>\n");
-            return RETURN_FAILURE;
-        }
-        const char *name =
-            data.cmdargtoken[2].val.string;
-        for(int i = 0; i < bookmark_count; i++)
-        {
-            if(strcmp(bookmarks[i].name,
-                     name) == 0)
-            {
-                strncpy(data.CLIcmdline,
-                        bookmarks[i].cmd,
-                        STRINGMAXLEN_CLICMDLINE
-                        - 1);
-                data.CLIcmdline[
-                    STRINGMAXLEN_CLICMDLINE - 1]
-                    = '\0';
-                return CLI_execute_line();
-            }
-        }
-        printf("Bookmark '%s' not found\n",
-               name);
-        return RETURN_FAILURE;
-    }
-
-    if(strcmp(action, "rm") == 0)
-    {
-        if(data.cmdNBarg < 3)
-        {
-            printf("Usage: bookmark rm "
-                   "<name>\n");
-            return RETURN_FAILURE;
-        }
-        const char *name =
-            data.cmdargtoken[2].val.string;
-        for(int i = 0; i < bookmark_count; i++)
-        {
-            if(strcmp(bookmarks[i].name,
-                     name) == 0)
-            {
-                for(int j = i;
-                        j < bookmark_count - 1;
-                        j++)
-                {
-                    bookmarks[j] =
-                        bookmarks[j + 1];
-                }
-                bookmark_count--;
-                cli_bookmark_save();
-                printf("Bookmark '%s' "
-                       "removed\n", name);
-                return RETURN_SUCCESS;
-            }
-        }
-        printf("Bookmark '%s' not found\n",
-               name);
-        return RETURN_FAILURE;
-    }
-
-    printf("Unknown bookmark action '%s'\n",
-           action);
-    return RETURN_FAILURE;
-}
-
-
-/*
- * ============================================================
- *  Session Logging
- * ============================================================
- */
-
-static FILE *session_log_fp = NULL;
-static struct timespec session_log_t0;
-
-errno_t cli_sessionlog(void)
-{
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: sessionlog "
-               "[on|off|<filename>]\n");
-        printf("Status: %s\n",
-               session_log_fp ? "ON" : "OFF");
-        return RETURN_SUCCESS;
-    }
-    const char *arg =
-        data.cmdargtoken[1].val.string;
-
-    if(strcmp(arg, "off") == 0)
-    {
-        if(session_log_fp)
-        {
-            fclose(session_log_fp);
-            session_log_fp = NULL;
-            printf("Session logging stopped\n");
-        }
-        return RETURN_SUCCESS;
-    }
-
-    /* Close previous if open */
-    if(session_log_fp)
-    {
-        fclose(session_log_fp);
-        session_log_fp = NULL;
-    }
-
-    char logpath[STRINGMAXLEN_FULLFILENAME];
-    if(strcmp(arg, "on") == 0)
-    {
-        snprintf(logpath,
-                 STRINGMAXLEN_FULLFILENAME,
-                 "%s/.milk_session.log",
-                 getenv("HOME"));
-    }
-    else
-    {
-        strncpy(logpath, arg,
-                STRINGMAXLEN_FULLFILENAME - 1);
-        logpath[
-            STRINGMAXLEN_FULLFILENAME - 1]
-            = '\0';
-    }
-
-    session_log_fp = fopen(logpath, "a");
-    if(session_log_fp == NULL)
-    {
-        printf("Cannot open '%s'\n", logpath);
-        return RETURN_FAILURE;
-    }
-    clock_gettime(CLOCK_MONOTONIC,
-                  &session_log_t0);
-    printf("Session logging to '%s'\n", logpath);
-
-    /* Write session start marker */
-    {
-        time_t now = time(NULL);
-        char tbuf[64];
-        strftime(tbuf, sizeof(tbuf),
-                 "%Y-%m-%dT%H:%M:%S",
-                 localtime(&now));
-        fprintf(session_log_fp,
-                "# Session started %s\n", tbuf);
-        fflush(session_log_fp);
-    }
-    return RETURN_SUCCESS;
-}
-
-/**
- * @brief Log a command to session log if active
- */
-static void cli_session_log_cmd(
-    const char *cmd
-)
-{
-    if(session_log_fp == NULL)
-    {
-        return;
-    }
-    struct timespec now;
-    clock_gettime(CLOCK_MONOTONIC, &now);
-    double elapsed_ms =
-        (double)(now.tv_sec
-                 - session_log_t0.tv_sec)
-        * 1000.0
-        + (double)(now.tv_nsec
-                   - session_log_t0.tv_nsec)
-        / 1.0e6;
-    {
-        time_t t = time(NULL);
-        char tbuf[64];
-        strftime(tbuf, sizeof(tbuf),
-                 "%Y-%m-%dT%H:%M:%S",
-                 localtime(&t));
-        fprintf(session_log_fp,
-                "[%s] [%10.1f ms] %s\n",
-                tbuf, elapsed_ms, cmd);
-        fflush(session_log_fp);
-    }
-}
-
-
-/*
- * ============================================================
- *  Brace Expansion
- * ============================================================
- *
- * Expand {N..M} into space-separated integers,
- * and {N..M..S} with step S.
- */
-
-/**
- * @brief Expand {N..M} and {N..M..S} brace ranges
- *
- * Replaces tokens like {1..5} with "1 2 3 4 5"
- * and {0..10..2} with "0 2 4 6 8 10".
- */
-static void emit_str(
-    char       *out,
-    int        *opos,
-    int         maxlen,
-    const char *s
-);
-static void cli_expand_braces(
-    char *line,
-    int   maxlen
-)
-{
-    char out[STRINGMAXLEN_CLICMDLINE];
-    int  opos = 0;
-    int  i = 0;
-
-    while(line[i] != '\0'
-          && opos < maxlen - 1)
-    {
-        if(line[i] == '{')
-        {
-            /* Try {N..M} or {N..M..S} */
-            char *endp = NULL;
-            long sv =
-                strtol(line + i + 1,
-                       &endp, 10);
-            if(endp != NULL
-               && endp[0] == '.'
-               && endp[1] == '.')
-            {
-                char *endp2 = NULL;
-                long ev =
-                    strtol(endp + 2,
-                           &endp2, 10);
-                long step = 1;
-                if(endp2 != NULL
-                   && endp2[0] == '.'
-                   && endp2[1] == '.')
-                {
-                    char *endp3 = NULL;
-                    step =
-                        strtol(endp2 + 2,
-                               &endp3,
-                               10);
-                    endp2 = endp3;
-                }
-                if(endp2 != NULL
-                   && *endp2 == '}'
-                   && step != 0)
-                {
-                    int first = 1;
-                    if(sv <= ev)
-                    {
-                        if(step < 0)
-                        {
-                            step = -step;
-                        }
-                        for(long v = sv;
-                            v <= ev;
-                            v += step)
-                        {
-                            char nb[32];
-                            snprintf(
-                                nb,
-                                sizeof(nb),
-                                "%s%ld",
-                                first
-                                ? "" : " ",
-                                v);
-                            first = 0;
-                            emit_str(
-                                out, &opos,
-                                maxlen, nb);
-                        }
-                    }
-                    else
-                    {
-                        if(step > 0)
-                        {
-                            step = -step;
-                        }
-                        for(long v = sv;
-                            v >= ev;
-                            v += step)
-                        {
-                            char nb[32];
-                            snprintf(
-                                nb,
-                                sizeof(nb),
-                                "%s%ld",
-                                first
-                                ? "" : " ",
-                                v);
-                            first = 0;
-                            emit_str(
-                                out, &opos,
-                                maxlen, nb);
-                        }
-                    }
-                    i = (int)(endp2
-                              - line) + 1;
-                    continue;
-                }
-            }
-            out[opos++] = line[i++];
-        }
-        else
-        {
-            out[opos++] = line[i++];
-        }
-    }
-    out[opos] = '\0';
-    strncpy(line, out, (size_t) maxlen);
-    line[maxlen - 1] = '\0';
-}
-
-
-/**
- * @brief Expand tilde (~) to $HOME
- *
- * Replaces ~ or ~/path at start of tokens
- * with the HOME environment variable value.
- */
-static void cli_expand_tilde(
-    char *line,
-    int   maxlen
-)
-{
-    const char *home = getenv("HOME");
-    if(home == NULL)
-    {
-        return;
-    }
-    char out[STRINGMAXLEN_CLICMDLINE];
-    int  opos = 0;
-    int  i = 0;
-    int  at_tok_start = 1;
-    int  in_sq = 0;
-    int  in_dq = 0;
-
-    while(line[i] != '\0'
-          && opos < maxlen - 1)
-    {
-        char c = line[i];
-        if(c == '\'' && !in_dq)
-        {
-            in_sq = !in_sq;
-            out[opos++] = line[i++];
-            at_tok_start = 0;
-            continue;
-        }
-        if(c == '"' && !in_sq)
-        {
-            in_dq = !in_dq;
-            out[opos++] = line[i++];
-            at_tok_start = 0;
-            continue;
-        }
-        if(c == ' ' || c == '\t')
-        {
-            out[opos++] = line[i++];
-            at_tok_start = 1;
-            continue;
-        }
-        if(at_tok_start
-           && !in_sq && !in_dq
-           && c == '~'
-           && (line[i + 1] == '/'
-               || line[i + 1] == ' '
-               || line[i + 1] == '\t'
-               || line[i + 1] == '\0'))
-        {
-            /* Replace ~ with $HOME */
-            const char *hp = home;
-            while(*hp != '\0'
-                  && opos < maxlen - 1)
-            {
-                out[opos++] = *hp++;
-            }
-            i++; /* skip ~ */
-            at_tok_start = 0;
-            continue;
-        }
-        out[opos++] = line[i++];
-        at_tok_start = 0;
-    }
-    out[opos] = '\0';
-    strncpy(line, out, (size_t) maxlen);
-    line[maxlen - 1] = '\0';
-}
-
-/**
- * @brief Expand filename globs (* and ?)
- *
- * Tokens containing * or ? that are not inside
- * quotes are expanded using POSIX glob().
- * Example: *.fits → file1.fits file2.fits
- */
-static void cli_expand_globs(
-    char *line,
-    int   maxlen
-)
-{
-    char out[STRINGMAXLEN_CLICMDLINE];
-    int  opos = 0;
-    int  i = 0;
-    int  in_sq = 0;
-    int  in_dq = 0;
-
-    while(line[i] != '\0'
-          && opos < maxlen - 1)
-    {
-        char c = line[i];
-        if(c == '\'' && !in_dq)
-        {
-            in_sq = !in_sq;
-            out[opos++] = line[i++];
-            continue;
-        }
-        if(c == '"' && !in_sq)
-        {
-            in_dq = !in_dq;
-            out[opos++] = line[i++];
-            continue;
-        }
-        if(in_sq || in_dq)
-        {
-            out[opos++] = line[i++];
-            continue;
-        }
-        if(c == ' ' || c == '\t')
-        {
-            out[opos++] = line[i++];
-            continue;
-        }
-        /* Extract token */
-        int tstart = i;
-        int has_glob = 0;
-        while(line[i] != '\0'
-              && line[i] != ' '
-              && line[i] != '\t')
-        {
-            if(line[i] == '*'
-               || line[i] == '?')
-            {
-                has_glob = 1;
-            }
-            i++;
-        }
-        int tlen = i - tstart;
-        if(!has_glob || tlen <= 0)
-        {
-            for(int j = tstart;
-                j < i
-                && opos < maxlen - 1;
-                j++)
-            {
-                out[opos++] = line[j];
-            }
-            continue;
-        }
-        /* Run glob */
-        char pat[512];
-        int plen = tlen;
-        if(plen >= 512)
-        {
-            plen = 511;
-        }
-        memcpy(pat, line + tstart,
-               (size_t) plen);
-        pat[plen] = '\0';
-
-        glob_t gl;
-        int gret = glob(pat,
-                        GLOB_NOCHECK,
-                        NULL, &gl);
-        if(gret == 0
-           && gl.gl_pathc > 0)
-        {
-            for(size_t g = 0;
-                g < gl.gl_pathc; g++)
-            {
-                if(g > 0
-                   && opos < maxlen - 1)
-                {
-                    out[opos++] = ' ';
-                }
-                const char *gp =
-                    gl.gl_pathv[g];
-                while(*gp != '\0'
-                      && opos
-                      < maxlen - 1)
-                {
-                    out[opos++] = *gp++;
-                }
-            }
-            globfree(&gl);
-        }
-        else
-        {
-            if(gret == 0)
-            {
-                globfree(&gl);
-            }
-            for(int j = tstart;
-                j < tstart + tlen
-                && opos < maxlen - 1;
-                j++)
-            {
-                out[opos++] = line[j];
-            }
-        }
-    }
-    out[opos] = '\0';
-    strncpy(line, out, (size_t) maxlen);
-    line[maxlen - 1] = '\0';
-}
-
-
-/*
- * ============================================================
- *  Command Substitution
- * ============================================================
- *
- * Replace $(cmd) and `cmd` in the command line with
- * the standard output of the command execution.
- */
-
-static void cli_expand_cmdsub(
-    char *line,
-    int   maxlen
-)
-{
-    char out[STRINGMAXLEN_CLICMDLINE];
-    int  opos = 0;
-    int  i = 0;
-
-    while(line[i] != '\0' && opos < maxlen - 1)
-    {
-        int is_dollar_paren =
-            (line[i] == '$'
-             && line[i + 1] == '('
-             && line[i + 2] != '(');
-        int is_backtick = (line[i] == '`');
-
-        if(is_dollar_paren || is_backtick)
-        {
-            char cmd[512];
-            int clen = 0;
-            
-            if (is_dollar_paren)
-            {
-                i += 2; /* Skip $( */
-                while(line[i] != '\0' && line[i] != ')' && clen < 511)
-                {
-                    cmd[clen++] = line[i++];
-                }
-                if (line[i] == ')') i++; /* Skip ) */
-            }
-            else /* is_backtick */
-            {
-                i++; /* Skip ` */
-                while(line[i] != '\0' && line[i] != '`' && clen < 511)
-                {
-                    cmd[clen++] = line[i++];
-                }
-                if (line[i] == '`') i++; /* Skip ` */
-            }
-            cmd[clen] = '\0';
-
-            /* Execute command and read output */
-            if (clen > 0)
-            {
-                FILE *fp = popen(cmd, "r");
-                if (fp != NULL)
-                {
-                    char buf[1024];
-                    size_t read_bytes = fread(buf, 1, sizeof(buf) - 1, fp);
-                    buf[read_bytes] = '\0';
-                    pclose(fp);
-
-                    /* Strip trailing newlines */
-                    while(read_bytes > 0 && (buf[read_bytes - 1] == '\n' || buf[read_bytes - 1] == '\r'))
-                    {
-                        buf[--read_bytes] = '\0';
-                    }
-
-                    /* Copy to output */
-                    int vallen = (int) read_bytes;
-                    int avail = maxlen - 1 - opos;
-                    int copylen = vallen < avail ? vallen : avail;
-                    memcpy(out + opos, buf, (size_t) copylen);
-                    opos += copylen;
-                }
-            }
-        }
-        else
-        {
-            out[opos++] = line[i++];
-        }
-    }
-    out[opos] = '\0';
-    strncpy(line, out, (size_t) maxlen);
-    line[maxlen - 1] = '\0';
-}
-
-
-/*
- * ============================================================
- *  Environment Variable Expansion
- * ============================================================
- *
- * Replace $VAR and ${VAR} with variable values.
- * Supports string operations, arrays, and special
- * forms inside ${...}.
- */
-
-/**
- * @brief Emit string into output buffer
- */
-static void emit_str(
-    char *out,
-    int  *opos,
-    int   maxlen,
-    const char *s
-)
-{
-    while(*s != '\0' && *opos < maxlen - 1)
-    {
-        out[(*opos)++] = *s++;
-    }
-}
-
-/**
- * @brief Handle ${...} braced expansion
- *
- * Supports:
- *   ${var}          plain lookup
- *   ${#var}         string length
- *   ${var:-default} default if unset
- *   ${var:=default} assign if unset
- *   ${var:+alt}     alt if set
- *   ${var:?error}   error if unset
- *   ${var:off:len}  substring
- *   ${var%%pat}     strip longest suffix
- *   ${var##pat}     strip longest prefix
- *   ${var%pat}      strip shortest suffix
- *   ${var#pat}      strip shortest prefix
- *   ${var/find/rep} replace first
- *   ${var//find/rep} replace all
- *   ${arr[N]}       array element
- *   ${arr[@]}       all array elements
- *   ${#arr[@]}      array element count
- */
-static void expand_braced(
-    char *out,
-    int  *opos,
-    int   maxlen,
-    const char *inner
-)
-{
-    /* ${#...} — length or array count */
-    if(inner[0] == '#')
-    {
-        const char *nm = inner + 1;
-        /* ${#arr[@]} */
-        const char *br = strchr(nm, '[');
-        if(br != NULL)
-        {
-            char aname[CLI_VAR_NAMELEN];
-            int alen = (int)(br - nm);
-            if(alen >= CLI_VAR_NAMELEN)
-            {
-                alen = CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(aname, nm,
-                   (size_t) alen);
-            aname[alen] = '\0';
-            for(int k = 0;
-                k < CLI_MAX_ARRAYS; k++)
-            {
-                if(cli_arrays[k].used
-                   && strcmp(
-                       cli_arrays[k].name,
-                       aname) == 0)
-                {
-                    char nb[32];
-                    snprintf(
-                        nb, sizeof(nb),
-                        "%d",
-                        cli_arrays[k]
-                        .nelem);
-                    emit_str(out, opos,
-                             maxlen, nb);
-                    return;
-                }
-            }
-            emit_str(out, opos,
-                     maxlen, "0");
-            return;
-        }
-        /* ${#var} — string length */
-        const char *val =
-            cli_var_lookup(nm);
-        char lb[32];
-        snprintf(lb, sizeof(lb), "%d",
-                 val ? (int) strlen(val)
-                 : 0);
-        emit_str(out, opos, maxlen, lb);
-        return;
-    }
-
-    /* ${!var} — indirect expansion */
-    if(inner[0] == '!')
-    {
-        const char *iname =
-            inner + 1;
-        const char *iref =
-            cli_var_lookup(iname);
-        if(iref != NULL)
-        {
-            const char *ival =
-                cli_var_lookup(iref);
-            if(ival != NULL)
-            {
-                emit_str(out, opos,
-                         maxlen,
-                         ival);
-            }
-        }
-        return;
-    }
-
-    /* ${arr[N]} or ${arr[@]} or
-     * ${assoc[key]} */
-    {
-        const char *br =
-            strchr(inner, '[');
-        if(br != NULL)
-        {
-            char aname[CLI_VAR_NAMELEN];
-            int alen = (int)(br - inner);
-            if(alen >= CLI_VAR_NAMELEN)
-            {
-                alen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(aname, inner,
-                   (size_t) alen);
-            aname[alen] = '\0';
-            const char *idx_s = br + 1;
-            if(idx_s[0] == '@')
-            {
-                for(int k = 0;
-                    k < CLI_MAX_ARRAYS;
-                    k++)
-                {
-                    if(cli_arrays[k].used
-                       && strcmp(
-                           cli_arrays[k]
-                           .name,
-                           aname) == 0)
-                    {
-                        for(int e = 0;
-                            e
-                            < cli_arrays[k]
-                            .nelem;
-                            e++)
-                        {
-                            if(e > 0)
-                            {
-                                emit_str(
-                                    out,
-                                    opos,
-                                    maxlen,
-                                    " ");
-                            }
-                            emit_str(
-                                out,
-                                opos,
-                                maxlen,
-                                cli_arrays
-                                [k]
-                                .elem[e]);
-                        }
-                        return;
-                    }
-                }
-                return;
-            }
-            int idx = (int) strtol(
-                idx_s, NULL, 0);
-            for(int k = 0;
-                k < CLI_MAX_ARRAYS;
-                k++)
-            {
-                if(cli_arrays[k].used
-                   && strcmp(
-                       cli_arrays[k]
-                       .name,
-                       aname) == 0)
-                {
-                    if(idx >= 0
-                       && idx
-                       < cli_arrays[k]
-                       .nelem)
-                    {
-                        emit_str(
-                            out, opos,
-                            maxlen,
-                            cli_arrays[k]
-                            .elem[idx]);
-                    }
-                    return;
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${assoc[key]} — associative
-     * array lookup */
-    {
-        const char *br =
-            strchr(inner, '[');
-        if(br != NULL)
-        {
-            const char *brend =
-                strchr(br, ']');
-            if(brend != NULL)
-            {
-                char aname[
-                    CLI_VAR_NAMELEN];
-                int nl =
-                    (int)(br
-                          - inner);
-                if(nl
-                   >= CLI_VAR_NAMELEN)
-                {
-                    nl =
-                        CLI_VAR_NAMELEN
-                        - 1;
-                }
-                memcpy(aname,
-                       inner,
-                       (size_t) nl);
-                aname[nl] = '\0';
-                char key[
-                    CLI_VAR_NAMELEN];
-                int kl =
-                    (int)(brend
-                          - br - 1);
-                if(kl
-                   >= CLI_VAR_NAMELEN)
-                {
-                    kl =
-                        CLI_VAR_NAMELEN
-                        - 1;
-                }
-                memcpy(key, br + 1,
-                       (size_t) kl);
-                key[kl] = '\0';
-                for(int k = 0;
-                    k
-                    < CLI_MAX_ASSOC;
-                    k++)
-                {
-                    if(cli_assoc[k]
-                        .used
-                       && strcmp(
-                           cli_assoc[
-                               k]
-                           .name,
-                           aname)
-                       == 0)
-                    {
-                        for(int e = 0;
-                            e
-                            < cli_assoc[
-                                k]
-                            .nelem;
-                            e++)
-                        {
-                            if(strcmp(
-                                cli_assoc[
-                                    k]
-                                .keys[e],
-                                key)
-                               == 0)
-                            {
-                                emit_str(
-                                    out,
-                                    opos,
-                                    maxlen,
-                                    cli_assoc[
-                                        k]
-                                    .vals[
-                                        e]);
-                                return;
-                            }
-                        }
-                        return;
-                    }
-                }
-            }
-        }
-    }
-
-    /* ${var//find/rep} — replace all */
-    {
-        const char *ds =
-            strstr(inner, "//");
-        if(ds != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen = (int)(ds - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *find =
-                ds + 2;
-            const char *sl2 =
-                strchr(find, '/');
-            char fp[256] = "";
-            char rp[256] = "";
-            if(sl2 != NULL)
-            {
-                int fl2 =
-                    (int)(sl2 - find);
-                if(fl2 > 255)
-                {
-                    fl2 = 255;
-                }
-                memcpy(fp, find,
-                       (size_t) fl2);
-                fp[fl2] = '\0';
-                strncpy(rp,
-                        sl2 + 1, 255);
-            }
-            else
-            {
-                strncpy(fp, find, 255);
-            }
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL
-               && fp[0] != '\0')
-            {
-                int fplen =
-                    (int) strlen(fp);
-                int rplen =
-                    (int) strlen(rp);
-                const char *s = val;
-                while(*s != '\0'
-                      && *opos
-                      < maxlen - 1)
-                {
-                    if(strncmp(s, fp,
-                               (size_t)
-                               fplen)
-                       == 0)
-                    {
-                        emit_str(
-                            out, opos,
-                            maxlen, rp);
-                        s += fplen;
-                    }
-                    else
-                    {
-                        out[(*opos)++] =
-                            *s++;
-                    }
-                }
-            }
-            else if(val != NULL)
-            {
-                emit_str(out, opos,
-                         maxlen, val);
-            }
-            return;
-        }
-    }
-
-    /* ${var/find/rep} — replace first */
-    {
-        const char *sl =
-            strchr(inner, '/');
-        if(sl != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen = (int)(sl - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *find = sl + 1;
-            const char *sl2 =
-                strchr(find, '/');
-            char fp[256] = "";
-            char rp[256] = "";
-            if(sl2 != NULL)
-            {
-                int fl2 =
-                    (int)(sl2 - find);
-                if(fl2 > 255)
-                {
-                    fl2 = 255;
-                }
-                memcpy(fp, find,
-                       (size_t) fl2);
-                fp[fl2] = '\0';
-                strncpy(rp,
-                        sl2 + 1, 255);
-            }
-            else
-            {
-                strncpy(fp, find, 255);
-            }
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL
-               && fp[0] != '\0')
-            {
-                const char *m =
-                    strstr(val, fp);
-                if(m != NULL)
-                {
-                    int pre =
-                        (int)(m - val);
-                    int avail =
-                        maxlen - 1
-                        - *opos;
-                    if(pre > avail)
-                    {
-                        pre = avail;
-                    }
-                    memcpy(
-                        out + *opos,
-                        val,
-                        (size_t) pre);
-                    *opos += pre;
-                    emit_str(
-                        out, opos,
-                        maxlen, rp);
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        m + strlen(fp));
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${var%%pattern} — strip suffix */
-    {
-        const char *pp =
-            strstr(inner, "%%");
-        if(pp != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(pp - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *pat = pp + 2;
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                const char *m =
-                    strstr(val, pat);
-                if(m != NULL)
-                {
-                    int clen =
-                        (int)(m - val);
-                    int avail =
-                        maxlen - 1
-                        - *opos;
-                    if(clen > avail)
-                    {
-                        clen = avail;
-                    }
-                    memcpy(
-                        out + *opos,
-                        val,
-                        (size_t) clen);
-                    *opos += clen;
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${var##pattern} — strip prefix */
-    {
-        const char *pp =
-            strstr(inner, "##");
-        if(pp != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(pp - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *pat = pp + 2;
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                const char *m =
-                    strstr(val, pat);
-                if(m != NULL)
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        m + strlen(pat));
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${var%pat} — strip shortest suffix */
-    {
-        const char *pp =
-            strchr(inner, '%');
-        if(pp != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(pp - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *pat = pp + 1;
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                int vl =
-                    (int) strlen(val);
-                int pl =
-                    (int) strlen(pat);
-                /* Find last occurrence */
-                const char *last =
-                    NULL;
-                const char *s = val;
-                while((s = strstr(
-                           s, pat))
-                      != NULL)
-                {
-                    last = s;
-                    s++;
-                }
-                if(last != NULL
-                   && (last + pl)
-                   == (val + vl))
-                {
-                    int clen =
-                        (int)(last
-                              - val);
-                    int avail =
-                        maxlen - 1
-                        - *opos;
-                    if(clen > avail)
-                    {
-                        clen = avail;
-                    }
-                    memcpy(
-                        out + *opos,
-                        val,
-                        (size_t) clen);
-                    *opos += clen;
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${var^^} uppercase / ${var,,} lowercase
-     * ${var^}  first char upper
-     * ${var,}  first char lower */
-    {
-        /* Find ^ or , in inner */
-        const char *cp =
-            strchr(inner, '^');
-        const char *cl =
-            strchr(inner, ',');
-        /* Pick the earlier one */
-        const char *op = NULL;
-        if(cp != NULL && cl != NULL)
-        {
-            op = (cp < cl) ? cp : cl;
-        }
-        else if(cp != NULL)
-        {
-            op = cp;
-        }
-        else if(cl != NULL)
-        {
-            op = cl;
-        }
-        if(op != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(op - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                char tmp[
-                    CLI_VAR_VALLEN];
-                strncpy(tmp, val,
-                        CLI_VAR_VALLEN
-                        - 1);
-                tmp[CLI_VAR_VALLEN
-                    - 1] = '\0';
-                if(op[0] == '^'
-                   && op[1] == '^')
-                {
-                    /* Uppercase all */
-                    for(int k = 0;
-                        tmp[k]
-                        != '\0'; k++)
-                    {
-                        tmp[k] =
-                            (char)
-                            toupper(
-                                (unsigned
-                                 char)
-                                tmp[k]);
-                    }
-                }
-                else if(op[0] == '^')
-                {
-                    /* First char */
-                    if(tmp[0] != '\0')
-                    {
-                        tmp[0] =
-                            (char)
-                            toupper(
-                                (unsigned
-                                 char)
-                                tmp[0]);
-                    }
-                }
-                else if(op[0] == ','
-                        && op[1]
-                        == ',')
-                {
-                    /* Lowercase all */
-                    for(int k = 0;
-                        tmp[k]
-                        != '\0'; k++)
-                    {
-                        tmp[k] =
-                            (char)
-                            tolower(
-                                (unsigned
-                                 char)
-                                tmp[k]);
-                    }
-                }
-                else if(op[0] == ',')
-                {
-                    /* First char */
-                    if(tmp[0] != '\0')
-                    {
-                        tmp[0] =
-                            (char)
-                            tolower(
-                                (unsigned
-                                 char)
-                                tmp[0]);
-                    }
-                }
-                emit_str(out, opos,
-                         maxlen,
-                         tmp);
-            }
-            return;
-        }
-    }
-
-    /* ${var#pat} — strip shortest prefix */
-    {
-        const char *pp =
-            strchr(inner, '#');
-        if(pp != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(pp - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            const char *pat = pp + 1;
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                int pl =
-                    (int) strlen(pat);
-                if(strncmp(val, pat,
-                           (size_t) pl)
-                   == 0)
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        val + pl);
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-            }
-            return;
-        }
-    }
-
-    /* ${var:-default} ${var:=default}
-     * ${var:+alt} ${var:?error}
-     * ${var:offset:length} */
-    {
-        const char *col =
-            strchr(inner, ':');
-        if(col != NULL)
-        {
-            char vn[CLI_VAR_NAMELEN];
-            int nlen =
-                (int)(col - inner);
-            if(nlen >= CLI_VAR_NAMELEN)
-            {
-                nlen =
-                    CLI_VAR_NAMELEN - 1;
-            }
-            memcpy(vn, inner,
-                   (size_t) nlen);
-            vn[nlen] = '\0';
-            char op = col[1];
-
-            /* ${var:-default} */
-            if(op == '-')
-            {
-                const char *val =
-                    cli_var_lookup(vn);
-                if(val != NULL
-                   && val[0] != '\0')
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        col + 2);
-                }
-                return;
-            }
-            /* ${var:=default} */
-            if(op == '=')
-            {
-                const char *val =
-                    cli_var_lookup(vn);
-                if(val != NULL
-                   && val[0] != '\0')
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-                else
-                {
-                    cli_var_set(
-                        vn, col + 2);
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        col + 2);
-                }
-                return;
-            }
-            /* ${var:+alt} */
-            if(op == '+')
-            {
-                const char *val =
-                    cli_var_lookup(vn);
-                if(val != NULL
-                   && val[0] != '\0')
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen,
-                        col + 2);
-                }
-                return;
-            }
-            /* ${var:?error} */
-            if(op == '?')
-            {
-                const char *val =
-                    cli_var_lookup(vn);
-                if(val == NULL
-                   || val[0] == '\0')
-                {
-                    fprintf(stderr,
-                            "%s: %s\n",
-                            vn,
-                            col + 2);
-                }
-                else
-                {
-                    emit_str(
-                        out, opos,
-                        maxlen, val);
-                }
-                return;
-            }
-
-            /* ${var:offset:length} */
-            int offset = (int) strtol(
-                col + 1, NULL, 0);
-            int slen = -1;
-            const char *c2 =
-                strchr(col + 1, ':');
-            if(c2 != NULL)
-            {
-                slen = (int) strtol(
-                    c2 + 1, NULL, 0);
-            }
-            const char *val =
-                cli_var_lookup(vn);
-            if(val != NULL)
-            {
-                int vl =
-                    (int) strlen(val);
-                if(offset < 0)
-                {
-                    offset =
-                        vl + offset;
-                }
-                if(offset < 0)
-                {
-                    offset = 0;
-                }
-                if(offset >= vl)
-                {
-                    return;
-                }
-                int rem =
-                    vl - offset;
-                if(slen < 0
-                   || slen > rem)
-                {
-                    slen = rem;
-                }
-                int avail =
-                    maxlen - 1
-                    - *opos;
-                if(slen > avail)
-                {
-                    slen = avail;
-                }
-                memcpy(
-                    out + *opos,
-                    val + offset,
-                    (size_t) slen);
-                *opos += slen;
-            }
-            return;
-        }
-    }
-
-    /* Plain ${var} */
-    const char *val =
-        cli_var_lookup(inner);
-    if(val != NULL)
-    {
-        emit_str(out, opos, maxlen,
-                 val);
-    }
-}
-
-/**
- * @brief Expand $VAR and ${VAR} in place
- *
- * Handles string ops, arrays, and special
- * forms inside ${...}.
- */
-void cli_expand_env(
-    char *line,
-    int   maxlen
-)
-{
-    char out[STRINGMAXLEN_CLICMDLINE];
-    int  opos = 0;
-    int  i = 0;
-
-    while(line[i] != '\0'
-            && opos < maxlen - 1)
-    {
-        if(line[i] == '`')
-        {
-            i++; /* skip ` */
-            char cmdsub[STRINGMAXLEN_CLICMDLINE];
-            int  clen = 0;
-            while(line[i] != '\0' && line[i] != '`' && clen < STRINGMAXLEN_CLICMDLINE - 1)
-            {
-                cmdsub[clen++] = line[i++];
-            }
-            if(line[i] == '`')
-            {
-                i++;
-            }
-            cmdsub[clen] = '\0';
-
-            FILE *fp = popen(cmdsub, "r");
-            if(fp)
-            {
-                char   resbuf[4096];
-                size_t bytes_read = fread(resbuf, 1, sizeof(resbuf) - 1, fp);
-                resbuf[bytes_read] = '\0';
-
-                while(bytes_read > 0 &&
-                      (resbuf[bytes_read - 1] == '\n' || resbuf[bytes_read - 1] == '\r'))
-                {
-                    resbuf[--bytes_read] = '\0';
-                }
-
-                for(size_t k = 0; k < bytes_read; k++)
-                {
-                    if(resbuf[k] == '\n' || resbuf[k] == '\r')
-                    {
-                        resbuf[k] = ' ';
-                    }
-                }
-
-                for(size_t k = 0; k < bytes_read && opos < maxlen - 1; k++)
-                {
-                    out[opos++] = resbuf[k];
-                }
-                pclose(fp);
-            }
-            continue;
-        }
-
-        if(line[i] == '$')
-        {
-            /* Skip $(( — arithmetic */
-            if(line[i + 1] == '('
-               && line[i + 2] == '(')
-            {
-                out[opos++] = line[i++];
-                continue;
-            }
-            /* Handle $( — command subst */
-            if(line[i + 1] == '(')
-            {
-                i += 2; /* skip $( */
-                char cmdsub[STRINGMAXLEN_CLICMDLINE];
-                int  clen = 0;
-                int  cdepth = 1;
-                while(line[i] != '\0' && clen < STRINGMAXLEN_CLICMDLINE - 1)
-                {
-                    if(line[i] == '(')
-                    {
-                        cdepth++;
-                    }
-                    else if(line[i] == ')')
-                    {
-                        cdepth--;
-                        if(cdepth == 0)
-                        {
-                            i++;
-                            break;
-                        }
-                    }
-                    cmdsub[clen++] = line[i++];
-                }
-                cmdsub[clen] = '\0';
-
-                FILE *fp = popen(cmdsub, "r");
-                if(fp)
-                {
-                    char   resbuf[4096];
-                    size_t bytes_read = fread(resbuf, 1, sizeof(resbuf) - 1, fp);
-                    resbuf[bytes_read] = '\0';
-
-                    /* Trim trailing newlines */
-                    while(bytes_read > 0 &&
-                          (resbuf[bytes_read - 1] == '\n' || resbuf[bytes_read - 1] == '\r'))
-                    {
-                        resbuf[--bytes_read] = '\0';
-                    }
-
-                    /* Replace internal newlines with space */
-                    for(size_t k = 0; k < bytes_read; k++)
-                    {
-                        if(resbuf[k] == '\n' || resbuf[k] == '\r')
-                        {
-                            resbuf[k] = ' ';
-                        }
-                    }
-
-                    for(size_t k = 0; k < bytes_read && opos < maxlen - 1; k++)
-                    {
-                        out[opos++] = resbuf[k];
-                    }
-                    pclose(fp);
-                }
-                continue;
-            }
-            i++;
-            if(line[i] == '{')
-            {
-                i++;
-                char inner[512];
-                int ilen = 0;
-                int depth = 1;
-                while(line[i] != '\0'
-                      && ilen < 511)
-                {
-                    if(line[i] == '{')
-                    {
-                        depth++;
-                    }
-                    if(line[i] == '}')
-                    {
-                        depth--;
-                        if(depth == 0)
-                        {
-                            i++;
-                            break;
-                        }
-                    }
-                    inner[ilen++] = line[i++];
-                }
-                inner[ilen] = '\0';
-                expand_braced(out, &opos,
-                              maxlen, inner);
-            }
-            else
-            {
-                /* $VAR — simple unbraced */
-                char varname[256];
-                int  vlen = 0;
-                while(line[i] != '\0'
-                      && vlen < 255)
-                {
-                    char c = line[i];
-                    if(!((c >= 'A'
-                          && c <= 'Z')
-                         || (c >= 'a'
-                             && c <= 'z')
-                         || (c >= '0'
-                             && c <= '9')
-                         || c == '_'
-                         || c == '?'))
-                    {
-                        break;
-                    }
-                    varname[vlen++] =
-                        line[i++];
-                    if(c == '?')
-                    {
-                        break;
-                    }
-                }
-                varname[vlen] = '\0';
-                const char *val =
-                    cli_var_lookup(varname);
-                if(val != NULL)
-                {
-                    emit_str(out, &opos,
-                             maxlen, val);
-                }
-            }
-        }
-        else
-        {
-            out[opos++] = line[i++];
-        }
-    }
-    out[opos] = '\0';
-    strncpy(line, out, (size_t) maxlen);
-    line[maxlen - 1] = '\0';
-}
-
-/*
- * ============================================================
- *  Built-in cd / pwd
- * ============================================================
- */
-
-errno_t cli_cd(void)
-{
-    const char *dir = getenv("HOME");
-    if(data.cmdNBarg >= 2)
-    {
-        dir = data.cmdargtoken[1].val.string;
-    }
-    if(dir != NULL)
-    {
-        if(chdir(dir) != 0)
-        {
-            printf("cd: %s: %s\n", dir, strerror(errno));
-            return RETURN_FAILURE;
-        }
-    }
-    return RETURN_SUCCESS;
-}
-
-errno_t cli_pwd(void)
-{
-    char cwd[1024];
-    if(getcwd(cwd, sizeof(cwd)) != NULL)
-    {
-        printf("%s\n", cwd);
-        return RETURN_SUCCESS;
-    }
-    else
-    {
-        perror("pwd");
-        return RETURN_FAILURE;
-    }
-}
-
-
-/*
- * ============================================================
- *  History <N> Command
- * ============================================================
- */
-
-errno_t cli_history_show(void)
-{
-#ifdef USE_READLINE
-    int n = 20;  /* default */
-    if(data.cmdNBarg >= 2)
-    {
-        n = atoi(
-            data.cmdargtoken[1].val.string);
-        if(n <= 0)
-        {
-            n = 20;
-        }
-    }
-
-    HIST_ENTRY **hlist = history_list();
-    if(hlist == NULL)
-    {
-        printf("No history\n");
-        return RETURN_SUCCESS;
-    }
-
-    int total = history_length;
-    int start = total - n;
-    if(start < 0)
-    {
-        start = 0;
-    }
-    for(int i = start; i < total; i++)
-    {
-        printf(" %4d  %s\n",
-               i + 1, hlist[i]->line);
-    }
-#else
-    printf("Readline not available\n");
-#endif
-    return RETURN_SUCCESS;
-}
-
-
-/*
- * ============================================================
- *  Fuzzy History Search (searchhist)
- * ============================================================
- *
- * Search history for entries containing a
- * substring. Shows all matches with index.
- */
-
-errno_t cli_searchhist(void)
-{
-#ifdef USE_READLINE
-    if(data.cmdNBarg < 2)
-    {
-        printf("Usage: searchhist <pattern>\n");
-        return RETURN_SUCCESS;
-    }
-    const char *pattern =
-        data.cmdargtoken[1].val.string;
-
-    HIST_ENTRY **hlist = history_list();
-    if(hlist == NULL)
-    {
-        printf("No history\n");
-        return RETURN_SUCCESS;
-    }
-
-    int total = history_length;
-    int found = 0;
-    for(int i = 0; i < total; i++)
-    {
-        if(strcasestr(hlist[i]->line,
-                      pattern) != NULL)
-        {
-            /* Highlight matching substring */
-            const char *pos =
-                strcasestr(hlist[i]->line,
-                           pattern);
-            int pre = (int)(pos
-                            - hlist[i]->line);
-            int plen = (int) strlen(pattern);
-            printf(" %4d  %.*s"
-                   "\033[1;33m%.*s\033[0m"
-                   "%s\n",
-                   i + 1,
-                   pre, hlist[i]->line,
-                   plen, pos,
-                   pos + plen);
-            found++;
-        }
-    }
-    if(found == 0)
-    {
-        printf("No history entries match"
-               " '%s'\n", pattern);
-    }
-    else
-    {
-        printf("(%d match%s)\n",
-               found,
-               found == 1 ? "" : "es");
-    }
-#else
-    printf("Readline not available\n");
-#endif
-    return RETURN_SUCCESS;
-}
 
 
 errno_t CLI_execute_line()
@@ -6805,8 +3188,8 @@ errno_t CLI_execute_line()
  * Set by CLI_redisplay when a suggestion is shown.
  * Consumed by accept_suggestion when Right Arrow is pressed.
  */
-static char *pending_suggestion = NULL;
-static int   pending_replace_len = 0;
+char *pending_suggestion = NULL;
+int   pending_replace_len = 0;
 
 /**
  * @brief Accept the inline suggestion on Right Arrow
@@ -6815,7 +3198,7 @@ static int   pending_replace_len = 0;
  * exists, insert it. Otherwise, fall through to normal
  * cursor-right movement.
  */
-static int accept_suggestion(
+int accept_suggestion(
     int count,
     int key
 )
@@ -6848,7 +3231,7 @@ static int accept_suggestion(
 /**
  * @brief Store the suggestion suffix for Right Arrow
  */
-static void set_pending_suggestion(
+void set_pending_suggestion(
     const char *text,
     int replace_len
 )
@@ -6869,7 +3252,7 @@ static void set_pending_suggestion(
  *
  * Returns -1 if no match.
  */
-static int find_command_match(const char *firstword)
+int find_command_match(const char *firstword)
 {
     for(uint32_t cmdi = 0; cmdi < data.NBcmd; cmdi++)
     {
@@ -6888,7 +3271,7 @@ static int find_command_match(const char *firstword)
  * Strips \001..\002 escape wrappers that readline
  * uses to mark non-printing characters.
  */
-static int visible_prompt_len(void)
+int visible_prompt_len(void)
 {
     const char *p = rl_display_prompt
                         ? rl_display_prompt
@@ -6920,7 +3303,7 @@ static int visible_prompt_len(void)
  * Returns max chars that can be printed after the
  * cursor without wrapping to the next terminal line.
  */
-static int get_ghost_budget(void)
+int get_ghost_budget(void)
 {
     struct winsize ws;
     int cols = 80; /* fallback */
@@ -6949,7 +3332,7 @@ static int get_ghost_budget(void)
  * in the given ANSI style, then resets style.
  * Returns number of visible chars printed.
  */
-static int print_ghost(
+int print_ghost(
     const char *style,
     const char *text,
     int budget
@@ -6971,9 +3354,9 @@ static int print_ghost(
 /**
  * @brief State for the reserved hint area
  */
-static int hint_area_active = 0;
-static int cached_term_rows = 0;
-static int cached_term_cols = 0;
+int hint_area_active = 0;
+int cached_term_rows = 0;
+int cached_term_cols = 0;
 
 /**
  * @brief Set up scroll region reserving bottom line
@@ -7058,7 +3441,7 @@ void CLI_cleanup_scroll_region(void)
  * Paints the reserved bottom line with the command
  * syntax when a known command is being typed.
  */
-static void update_hint_area(void)
+void update_hint_area(void)
 {
     if(!hint_area_active || !data.autocomplete_arghint)
     {
@@ -7237,7 +3620,7 @@ static void update_hint_area(void)
     fflush(stdout);
 }
 
-static void CLI_redisplay(void)
+void CLI_redisplay(void)
 {
     /* Default or syntax-highlighted redisplay */
     rl_redisplay_function = NULL;
