+++ src/cli/CLIcore/CLIcore/CLIcore_UI.h
@@ -98,4 +98,20 @@ void    cli_history_log_init(void);
 errno_t cli_ghistory(void);
 errno_t cli_lhistory(void);
 
-#endif
+#define BOOKMARK_MAX      64
+#define BOOKMARK_NAMELEN  200
+#define BOOKMARK_CMDLEN  2000
+
+void cli_history_expand(void);
+void cli_alias_expand(void);
+void cli_expand_cmdsub(char *line, int maxlen);
+void cli_expand_tilde(char *line, int maxlen);
+void cli_expand_braces(char *line, int maxlen);
+void cli_expand_globs(char *line, int maxlen);
+void cli_session_log_cmd(const char *cmd);
+void cli_history_log_cmd(const char *cmd);
+const char *CLI_history_file(void);
+void cli_save_last_argument(void);
+const char *strip_ws(const char *s);
+
+#endif /* CLICORE_UI_H */
