+++ src/cli/CLIcore/CLIcore/ui_includes.h
@@ -1,0 +1,18 @@
+#include <stdio.h>
+#include <dirent.h>
+#include <sys/stat.h>
+#include <sys/ioctl.h>
+#include <unistd.h>
+#include <termios.h>
+#ifdef USE_READLINE
+#include <readline/history.h>
+#include <readline/readline.h>
+#endif
+#include "CLIcore.h"
+#include "CLIcore_UI.h"
+#include "CLIcore_script.h"
+#include "CLIcore/cli_calc_parser.h"
+#include <glob.h>
+#include <sys/wait.h>
+#include "COREMOD_memory/COREMOD_memory.h"
+#include "timeutils.h"
